refactor: HintViewController에서 PaddingLabel 분리
